chore: tidy the componets more meanful
